pyproject.toml
[project]
name = "trello-sankey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "mypy>=1.0.0"
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["trello_sankey"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]  # Ignore line length in test files

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "types-requests>=2.32.4.20250611",
]
